+++ packages/article-summary/src/styles/index.web.js
@@ -1,11 +1,12 @@
+import { StyleSheet } from "react-native";
 import sharedStyles from "./shared";
 
-const styles = {
+const styles = StyleSheet.create({
   ...sharedStyles,
   headline: {
     ...sharedStyles.headline,
     fontWeight: "400"
   }
-};
+});
 
 export default styles;