+++ packages/pages/src/article.js
@@ -4,14 +4,14 @@ import { ArticleProvider } from "@times-components/provider";
 import withClient from "./client/with-client";
 import SimpleArticle from "./simple-article";
 
-const ArticleDetailsPage = ({ articleId, ...props }) => (
+const ArticleDetailsPage = ({ articleId, omitErrors, ...props }) => (
   <ArticleProvider debounceTimeMs={100} id={articleId}>
     {({ article, isLoading, error, refetch }) => (
       <SimpleArticle
         {...props}
         article={article}
-        error={error}
-        isLoading={isLoading}
+        error={omitErrors ? null : error}
+        isLoading={isLoading || (omitErrors && error !== null)}
         refetch={refetch}
       />
     )}
@@ -22,6 +22,7 @@ ArticleDetailsPage.propTypes = {
   articleId: PropTypes.string.isRequired,
   analyticsStream: PropTypes.func.isRequired,
   platformAdConfig: PropTypes.shape({}).isRequired,
+  omitErrors: PropTypes.bool,
   onArticlePress: PropTypes.func.isRequired,
   onAuthorPress: PropTypes.func.isRequired,
   onCommentsPress: PropTypes.func.isRequired,
@@ -31,6 +32,10 @@ ArticleDetailsPage.propTypes = {
   onTopicPress: PropTypes.func.isRequired,
   scale: PropTypes.string.isRequired,
   sectionName: PropTypes.string.isRequired
+};
+
+ArticleDetailsPage.defaultProps = {
+  omitErrors: false
 };
 
 export default withClient(ArticleDetailsPage);