+++ packages/ssr/webpack.config.prod.js
@@ -1,0 +1,66 @@
+const path = require("path");
+const webpack = require("webpack");
+
+const alias = { "react-native": "react-native-web" };
+const extensions = [".web.js", ".js"];
+const mode = "production";
+
+const babelConfig = {
+  test: /\.js$/,
+  use: {
+    loader: "babel-loader",
+    options: {
+      cacheDirectory: true,
+      plugins: [
+        "add-react-displayname",
+        "react-native-web",
+        "styled-components"
+      ],
+      presets: ["react-native"]
+    }
+  }
+};
+
+if (!process.env.GRAPHQL_ENDPOINT) {
+  throw new Error("You need to set a GRAPHQL_ENDPOINT envar");
+}
+
+module.exports = {
+  devtool: "source-map",
+  entry: {
+    "author-profile": "./author-profile.client.js",
+    topic: "./topic.client.js"
+  },
+  mode,
+  module: {
+    rules: [babelConfig]
+  },
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        commons: {
+          name: "vendor",
+          chunks: "initial",
+          minChunks: 2,
+          minSize: 0
+        }
+      }
+    },
+    occurrenceOrder: true
+  },
+  output: {
+    filename: "[name].bundle.js",
+    path: path.resolve(__dirname, "./dist")
+  },
+  plugins: [
+    new webpack.DefinePlugin({
+      "process.env": {
+        GRAPHQL_ENDPOINT: JSON.stringify(process.env.GRAPHQL_ENDPOINT)
+      }
+    })
+  ],
+  resolve: {
+    alias,
+    extensions
+  }
+};