+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.2.12"></a>
+## [2.2.12](https://github.com/newsuk/times-components/compare/@times-components/icons@2.2.11...@times-components/icons@2.2.12) (2018-08-29)
+
+**Note:** Version bump only for package @times-components/icons
+
+
+
+
+
 <a name="2.2.11"></a>
 ## [2.2.11](https://github.com/newsuk/times-components/compare/@times-components/icons@2.2.10...@times-components/icons@2.2.11) (2018-08-24)