+++ packages/jest-configurator/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.5"></a>
+## [2.1.5](https://github.com/newsuk/times-components/compare/@times-components/jest-configurator@2.1.4...@times-components/jest-configurator@2.1.5) (2018-08-29)
+
+**Note:** Version bump only for package @times-components/jest-configurator
+
+
+
+
+
 <a name="2.1.4"></a>
 ## [2.1.4](https://github.com/newsuk/times-components/compare/@times-components/jest-configurator@2.1.3...@times-components/jest-configurator@2.1.4) (2018-08-13)