+++ packages/topic/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.3.0"></a>
+# [3.3.0](https://github.com/newsuk/times-components/compare/@times-components/topic@3.2.0...@times-components/topic@3.3.0) (2018-08-29)
+
+
+### Features
+
+* add ssr package ([#1273](https://github.com/newsuk/times-components/issues/1273)) ([11ae083](https://github.com/newsuk/times-components/commit/11ae083))
+
+
+
+
+
+<a name="3.2.0"></a>
+# [3.2.0](https://github.com/newsuk/times-components/compare/@times-components/topic@3.1.19...@times-components/topic@3.2.0) (2018-08-29)
+
+
+### Features
+
+* add markup support for `emphasis` and `strong`. ([#1271](https://github.com/newsuk/times-components/issues/1271)) ([c79a734](https://github.com/newsuk/times-components/commit/c79a734))
+
+
+
+
+
 <a name="3.1.19"></a>
 ## [3.1.19](https://github.com/newsuk/times-components/compare/@times-components/topic@3.1.18...@times-components/topic@3.1.19) (2018-08-24)